requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
pydantic
Authlib
PyJWT
